fix(git): keep archive snapshot authoritative until resume writes data
Addresses PR #2851 review comments (Codex P1, CodeRabbit, Claude,
Greptile):

- Codex P1: unarchive-before-resume regressed to pre-archive state.
  httpUnarchiveTask / UnarchiveTaskTree only clear archived_at and
  restore branches/workspaces; no snapshot is written until the
  session resumes and the agent completes another turn. The archive
  row now keeps rank-0 while the task is archived OR while no
  non-archive snapshot newer than it exists; once a newer
  agent_completed/live_monitor row lands, the stale archive correctly
  demotes. Covered in both selectors by the shared snapshotRankExpr,
  with a new unarchive-before-resume regression (single + batch).
- CodeRabbit: the CASE ranking + tie-breakers were duplicated across
  both selectors and could silently diverge; extracted into the
  package-level snapshotRankExpr constant interpolated by both.
- Claude: TestCreateGitSnapshotRoundTripsEveryField now archives the
  task so GetLatestGitSnapshot exercises the rank-0 branch rather
  than selecting the only row by construction.
- Greptile: moved the new special-path/section-boundary parser tests
  into git_log_diffstats_special_test.go (backend convention: new
  tests in a new file, not appended to an already-large test file).

Verified: repository + sqlite ordering suites, parser suite
(including moved tests), gofmt/vet, and golangci-lint --new-from-rev
all clean.
