+++ apps/backend/internal/task/repository/sqlite/git_snapshots.go
@@ -17,6 +17,46 @@ import (
 // UpsertLatestLiveGitSnapshot so we don't disturb archive/completion snapshots.
 const TriggeredByLiveMonitor = "live_monitor"
 
+// snapshotRankExpr is the ORDER BY ranking shared by GetLatestGitSnapshot and
+// GetLatestGitSnapshotsBySessionIDs. Both selectors must pick the same row for
+// a session, so the expression lives in one place:
+//
+//   - rank 0: an archive snapshot that is still authoritative — either its
+//     owning task is still archived (terminal cumulative diff), or the task
+//     was unarchived but not yet resumed and no non-archive snapshot newer
+//     than this archive row exists yet (a reload in that window must not
+//     regress to pre-archive state).
+//   - rank 1: agent_completed snapshots (captured at exact completion time).
+//   - rank 2: live_monitor snapshots (periodic polls; may race completion).
+//   - rank 3: archive snapshots that are stale (task unarchived AND a newer
+//     non-archive snapshot exists — resumed execution has moved on).
+//
+// Ties within a rank break by created_at DESC, then id DESC so the selection
+// is deterministic even when Postgres stores multiple writes at the same
+// microsecond timestamp.
+const snapshotRankExpr = `
+			CASE
+				WHEN snapshot_type = 'archive' AND (
+					EXISTS (
+						SELECT 1 FROM task_sessions ts
+						JOIN tasks t ON t.id = ts.task_id
+						WHERE ts.id = task_session_git_snapshots.session_id
+						  AND t.archived_at IS NOT NULL
+					)
+					OR NOT EXISTS (
+						SELECT 1 FROM task_session_git_snapshots newer
+						WHERE newer.session_id = task_session_git_snapshots.session_id
+						  AND newer.snapshot_type <> 'archive'
+						  AND newer.created_at > task_session_git_snapshots.created_at
+					)
+				) THEN 0
+				WHEN triggered_by = 'agent_completed' THEN 1
+				WHEN snapshot_type = 'archive' THEN 3
+				ELSE 2
+			END,
+			created_at DESC,
+			id DESC`
+
 // UpsertLatestLiveGitSnapshot keeps at most one cached "live monitor" snapshot
 // per session by deleting any previous live row and inserting the new one in a
 // single transaction. This is the cache that backs the sidebar diff badge for
@@ -163,15 +203,10 @@ func (r *Repository) getGitSnapshotByOrder(ctx context.Context, sessionID, order
 
 // GetLatestGitSnapshot retrieves the best git snapshot for a session.
 // An archive snapshot (terminal cumulative diff captured at archive time) is
-// authoritative only while its owning task is still archived; once the task
-// is unarchived and its session resumes, the archive row is stale and newer
-// agent_completed / live_monitor rows describe the current execution.
-// agent_completed (captured at exact completion time) is preferred over
-// live_monitor (periodic polls that may contain stale data if the poll raced
-// with agent completion). Equal-rank rows break ties by created_at DESC,
-// then id DESC so the selection is deterministic even when Postgres stores
-// multiple writes at the same microsecond timestamp. Falls back to
-// most-recent-by-time when none of those exist.
+// authoritative while its owning task is still archived, and also while the
+// task is unarchived but not yet resumed (no newer non-archive snapshot
+// exists yet); once a newer agent_completed / live_monitor row describes the
+// current execution, it wins over the stale archive row.
 // Returns sql.ErrNoRows if no snapshot is found.
 func (r *Repository) GetLatestGitSnapshot(ctx context.Context, sessionID string) (*models.GitSnapshot, error) {
 	query := `
@@ -179,20 +214,7 @@ func (r *Repository) GetLatestGitSnapshot(ctx context.Context, sessionID string)
 		       ahead, behind, files, triggered_by, metadata, created_at
 		FROM task_session_git_snapshots
 		WHERE session_id = ?
-		ORDER BY
-			CASE
-				WHEN snapshot_type = 'archive' AND EXISTS (
-					SELECT 1 FROM task_sessions ts
-					JOIN tasks t ON t.id = ts.task_id
-					WHERE ts.id = task_session_git_snapshots.session_id
-					  AND t.archived_at IS NOT NULL
-				) THEN 0
-				WHEN triggered_by = 'agent_completed' THEN 1
-				WHEN snapshot_type = 'archive' THEN 3
-				ELSE 2
-			END,
-			created_at DESC,
-			id DESC
+		ORDER BY` + snapshotRankExpr + `
 		LIMIT 1
 	`
 	return scanGitSnapshot(r.ro.QueryRowContext(ctx, r.ro.Rebind(query), sessionID))
@@ -200,10 +222,8 @@ func (r *Repository) GetLatestGitSnapshot(ctx context.Context, sessionID string)
 
 // GetLatestGitSnapshotsBySessionIDs loads one authoritative snapshot per
 // session in one query per placeholder-sized chunk. It mirrors
-// GetLatestGitSnapshot's archive(while archived) > agent_completed >
-// live_monitor > archive(stale) preference (with the same created_at DESC,
-// id DESC tie-break) without introducing an N+1 read when task-list
-// summaries repair historical rows.
+// GetLatestGitSnapshot's ranking (snapshotRankExpr) without introducing an
+// N+1 read when task-list summaries repair historical rows.
 func (r *Repository) GetLatestGitSnapshotsBySessionIDs(
 	ctx context.Context,
 	sessionIDs []string,
@@ -222,20 +242,7 @@ func (r *Repository) GetLatestGitSnapshotsBySessionIDs(
 				       ahead, behind, files, triggered_by, metadata, created_at,
 				       ROW_NUMBER() OVER (
 					       PARTITION BY session_id
-					       ORDER BY
-					               CASE
-						               WHEN snapshot_type = 'archive' AND EXISTS (
-							               SELECT 1 FROM task_sessions ts
-							               JOIN tasks t ON t.id = ts.task_id
-							               WHERE ts.id = task_session_git_snapshots.session_id
-							                 AND t.archived_at IS NOT NULL
-						               ) THEN 0
-						               WHEN triggered_by = 'agent_completed' THEN 1
-						               WHEN snapshot_type = 'archive' THEN 3
-						               ELSE 2
-					               END,
-					               created_at DESC,
-					               id DESC
+					       ORDER BY` + snapshotRankExpr + `
 				       ) AS row_number
 				FROM task_session_git_snapshots
 				WHERE session_id IN (` + placeholders + `)
